dom/imptests/webapps/WebStorage/tests/submissions/Infraware/test_storage_local_setitem.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>storage_local_setItem</h1>
    <div id="log"></div>
    <script>
        test(function() {
            localStorage.clear();
            localStorage.setItem("name", "user1");
            assert_equals(localStorage.length, 1, "localStorage.setItem")
        });
    </script>
 </body>
</html>
